+++ meta/defined.rb
@@ -7,3 +7,13 @@ Mutant::Meta::Example.add :defined? do
   mutation 'defined?(nil)'
   mutation 'true'
 end
+
+Mutant::Meta::Example.add :defined? do
+  source 'defined?(@foo)'
+
+  singleton_mutations
+  mutation 'defined?(nil)'
+  mutation 'defined?(foo)'
+  mutation 'true'
+  mutation 'instance_variable_defined?(:@foo)'
+end